+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Fixes:
 - Improved the `markdownEscape` function!
 - Translated even more MadelineProto UI elements!
 - Now the same static analysis rules are applied for both plugins and simple event handlers.
+- Made some fixes to simple filters.
 
 ---